lambda-ocr-final/requirements.txt
requests

pytesseract
Pillow
openpyxl
pdf2image==1.16.3
boto3
